chore: remove useless code
