--- examples/rax-kitchen-sink/src/components/App/index.css
@@ -1,27 +1,0 @@
-.app {
-  flex: 1;
-  justify-content: center;
-  align-items: center;
-  background-color: #222;
-  color: white;
-  padding: 40px;
-}
-
-.appHeader {
-  height: 160;
-  padding: 40px;
-}
-
-.appBanner, .appIntro {
-  text-align: center;
-}
-
-.appBanner {
-  font-size: 40;
-  color: white;
-}
-
-.appIntro {
-  margin-top: 40;
-  font-size: 20;
-}
