--- examples/rax-kitchen-sink/src/components/Welcome/Welcome.test.js
@@ -1,9 +1,0 @@
-import { createElement } from 'rax';
-import renderer from 'rax-test-renderer';
-import { Welcome } from './index';
-
-test('Welcome render correctly', () => {
-  const tree = renderer.create(<Welcome />).toJSON();
-
-  expect(tree).toMatchSnapshot();
-});
